Add LiteLLM proxy and logging modules for enhanced observability and request handling
